roachprod: move promethius package to cmd/roachprod
This simple refactor is prerequisite for adding a simple command that spins up
promethius/grafana instances for a roachprod cluster.

Release note: None
