+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -159,9 +159,9 @@ go_library(
         "//pkg/cloud/gcp",
         "//pkg/clusterversion",
         "//pkg/cmd/cmpconn",
+        "//pkg/cmd/roachprod/prometheus",
         "//pkg/cmd/roachtest/cluster",
         "//pkg/cmd/roachtest/option",
-        "//pkg/cmd/roachtest/prometheus",
         "//pkg/cmd/roachtest/registry",
         "//pkg/cmd/roachtest/spec",
         "//pkg/cmd/roachtest/test",
@@ -239,8 +239,8 @@ go_test(
     ],
     embed = [":tests"],
     deps = [
+        "//pkg/cmd/roachprod/prometheus",
         "//pkg/cmd/roachtest/option",
-        "//pkg/cmd/roachtest/prometheus",
         "//pkg/cmd/roachtest/spec",
         "//pkg/roachprod/logger",
         "//pkg/testutils/skip",
